+++ R/cut_tour.R
@@ -31,7 +31,6 @@
 #' @author Michael Hahsler
 #' @keywords optimize
 #' @examples
-#'
 #' data("USCA50")
 #'
 #' ## find a path starting at Austin, TX
@@ -53,43 +52,61 @@
 #' path
 #' @export
 cut_tour <- function(x, cut, exclude_cut = TRUE)
-    UseMethod("cut_tour")
+  UseMethod("cut_tour")
 
 #' @rdname cut_tour
 #' @export
 cut_tour.TOUR <- function(x, cut, exclude_cut = TRUE) {
+  exclude_cut <- as.integer(exclude_cut)
 
-    exclude_cut <- as.integer(exclude_cut)
-
-    ## city label
-    if(is.character(cut)) {
-        cut <- which(as.logical(apply(sapply(cut, "==", labels(x)), MARGIN = 1, sum)))
-        if(length(cut) < 1) stop("cut has to exist")
+  ## city label
+  if (is.character(cut)) {
+    cut <-
+      which(as.logical(apply(
+        sapply(cut, "==", labels(x)), MARGIN = 1, sum
+      )))
+    if (length(cut) < 1)
+      stop("cut has to exist")
 
     ## city id
-    } else {
-        if(any(is.na(cut <- match(cut, x)))) stop("cut has to exist")
+  } else {
+    if (any(is.na(cut <- match(cut, x))))
+      stop("cut has to exist")
+  }
+
+  if (length(cut) == 1L) {
+    ## single path
+    if (exclude_cut && length(x) <= 1)
+      path <- integer(0)
+    else
+      path <-
+        c(x, x)[(cut + exclude_cut):(length(x) + cut - 1L)]
+
+  } else {
+    ## multiple paths, return as a list
+    path <- replicate(length(cut), integer(0))
+    path[[1L]] <- c(if ((tail(cut, 1) + exclude_cut) <= length(x))
+      (tail(cut, 1) + exclude_cut):length(x)
+      else
+        NULL,
+      if (cut[1] > 1)
+        1:(cut[1] - 1L)
+      else
+        NULL)
+    for (i in seq_len(length(cut) - 1L)) {
+      if ((cut[i] + exclude_cut) <= (cut[i + 1L] - 1L))
+        path[[i + 1L]] <-
+          (cut[i] + exclude_cut):(cut[i + 1L] - 1L)
     }
 
-    if(length(cut) == 1L) { ## single path
-        if (exclude_cut && length(x) <= 1)
-            path <- integer(0)
-        else
-            path <- c(x, x)[(cut + exclude_cut):(length(x) + cut - 1L)]
+    path <- lapply(
+      path,
+      FUN = function(i)
+        x[i]
+    )
+    if (exclude_cut)
+      names(path) <- labels(x)[cut]
+  }
 
-    } else { ## multiple paths, return as a list
-        path <- replicate(length(cut), integer(0))
-        path[[1L]] <- c(
-            if ((tail(cut, 1) + exclude_cut) <= length(x)) (tail(cut, 1) + exclude_cut):length(x) else NULL,
-            if (cut[1] > 1) 1:(cut[1] - 1L) else NULL)
-        for (i in seq_len(length(cut) - 1L)) {
-            if ((cut[i] + exclude_cut) <= (cut[i + 1L] - 1L))
-                path[[i + 1L]] <- (cut[i] + exclude_cut):(cut[i + 1L] - 1L)
-        }
-
-        path <- lapply(path, FUN = function(i) x[i])
-        if (exclude_cut) names(path) <- labels(x)[cut]
-    }
-
-    path
+  path
 }