+++ tests/testthat/test-insert_cut_etc.R
@@ -66,6 +66,11 @@ context("reformulate")
 atsp <- as.ATSP(USCA50)
 expect_equal(tour_length(USCA50), tour_length(atsp))
 
-tsp <- reformulate_ATSP_as_TSP(atsp)
+tsp <- reformulate_ATSP_as_TSP(atsp, cheap = 0)
 expect_equal(n_of_cities(atsp)*2, n_of_cities(tsp))
 
+tour_tsp <- solve_TSP(tsp)
+tour_atsp <- filter_ATSP_as_TSP_dummies(tour_tsp, atsp)
+
+expect_equal(length(tour_atsp), n_of_cities(USCA50))
+expect_equal(tour_length(tour_tsp), tour_length(tour_atsp))